+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Cynosure `@VERSION@` `1.20.1` `Fabric & Forge`
 
 ### Changelog:
-- New Entity Render Layer registration allowing for abstract renderers
+- Fixed crash with renderer change we did